Ensure syntax options/configuration are respected in __str__
Previously we had problems when converting an Option with None value to
str, as reported in #68.
The changes introduced here try to "harmonise" the syntax options given
to the ConfigUpdater object and the behaviour of Option.__str__.

During the implementation of these changes, it was also noticed that
Section.set (and therefore Document.set) where not honouring these
settings either (i.e. `option._space_around_delimiters == True` even
when instantiating ConfigUpdater with `space_around_delimiters=False`).

A side effect of the implemented changes is the introduction of a
read-only property `syntax_options` for ConfigUpdater and Parser objects.