Add a diff tool to assist comparing changes to stdlib
Since ConfigUpdater is initially derived from the stdlib and
the stdlib is a moving target, from times to times it is nice to see if
there was any improvement that we are not adopting.

This is an attempt to add a tool that will make that job easier for us.